Create function to prompt user for a choice
Used a while loop to keep asking until the user types in one of the
three options or cancels. Not sure what will happen to the game if the
user cancels, this might not have been needed, but the instructions
aren't super clear. I'm thinking I could return 'null' or some other
value that would cause the computer to win, or just cause the game to
end.