Create a function to play one round of the game
Used nested control flow - if/else inside a switch inside an if/else. I
feel like there is a smoother way to do this, but it's something I'll
hold of on trying to do until later. Could be a good exercise. Had an
issue with the if/else statements in the switch, as I was assigning
not comparing. Used the debugger to suss out that issue, which, while
it feels like a rookie mistake, I'm proud I was able to step through
and see what my code was actually doing.