Add playGame function to play mulitple rounds
My initial version of the playRound function output the choices and
winner to the console log, as well as updating the score. playGame
calls this function. I could have left the console logs in playRound,
but I think it makes more sense for that function to simply return the
result to the playGame function, which can then log the results and
update the scores.