Add buttons to HTML file to start game, add CSS
Added a div with three buttons for the user to choose their weapon and
start each round. I used event delegation so that I only had to add an
event listener to the parent div, and set a variable to track the
button. Downside of this is that clicking anywhere in the div will
trigger the function call. I added CSS to minimize the width of the
div, and I am sure there are ways to prevent this from triggering the
playRound() function. Two things I can think of off the top of my head:

1. Code to "catch" the undefined click
2. Make the buttons take up all of the space in the div, so that the
user can't click anywhere except on a button