Add function to determine winner, div for results
I want playGame() to basically just call playRound() and then check if
the game should continue. Similarly, though, I want playRound to just
call the necessary functions and just format the round winner message.
determineRoundWinner() is basically just the if/else and switch
statements that run through the options and determine a winner.

I also declared the point in the main script, passing them into the
play function. This might change when I finally work on playGame(), and
they can be declared there, and then passed to playRound().