Add code to playRound() to increment scores
I could have used an if-else loop to check which player won, or if it
was a tie. Instead, I add the boolean result of whether the winner
matches the player. If the human won the round, the boolean will
resolve to 1. If they lost, it will be 0. The one hitch here is that I
account for ties, and award a point each, so, I had to include an OR
statement to account for ties. If I decided that ties would be 0 points
each, I could just leave that out. If I did something silly and made
ties worth 0.5 points, I'd have to figure something else out.