Edit the playGame() function to run a 5-loop game
I cluttered it up playGame() with if-else statements to determine the
end of game. I think I will move that into a separate function. I also
don't like that I'm returning an array from playRound(). This is legal,
but it feels clunky, and the resulting output has a lot of bracket
notation, which I feel makes things less clear. I would like to change
that a bit.