Modify play func's add victoryMessage()&gameOver()
I realized that determineRoundWinner() was unecessary, and it was what
playRound() should have been. playRound() is called from playGame() and
returns the winner. That's all that is needed to process the rest of
the game: winner determines the scores and the messaging, so playGame
can then either call functions or modify variables as needed.

I moved the code for setting the victory message as well as the code
for determining whether the game should continue or not out of the
playGame() function. victoryMessage() probably isn't technically
needed, but I want to keep my functions as clean as possible. I feel
like separating out the game over functionality makes a little more
sense.